リソースを削除した
