+++ CharacterAlarm/Entity/Resource.swift
@@ -1,21 +1,21 @@
-import Foundation
-
-struct ImageList: Codable {
-    let images: [String]
-}
-
-struct VoiceList: Codable {
-    let voices: [String]
-}
-
-struct ImageAndVoiceList: Codable {
-    let images: [String]
-    let voices: [String]
-}
-
-struct Resource: Codable {
-    let version: Int
-    let resource: ImageAndVoiceList
-    let expression: [String: ImageAndVoiceList]
-    let call: [String: VoiceList]
-}
+//import Foundation
+//
+//struct ImageList: Codable {
+//    let images: [String]
+//}
+//
+//struct VoiceList: Codable {
+//    let voices: [String]
+//}
+//
+//struct ImageAndVoiceList: Codable {
+//    let images: [String]
+//    let voices: [String]
+//}
+//
+//struct Resource: Codable {
+//    let version: Int
+//    let resource: ImageAndVoiceList
+//    let expression: [String: ImageAndVoiceList]
+//    let call: [String: VoiceList]
+//}
