+++ CharacterAlarm/Repository/ResourceStore.swift
@@ -17,15 +17,15 @@ class ResourceStore {
         return data
     }
     
-    static func downloadResourceJson(charaDomain: String) async throws -> Resource {
-        let path = "/\(charaDomain)/resource.json"
-        let url = URL(string: environmentVariable.apiEndpoint + path)!
-        let data = try await APIClient<String>().fileDownload(url: url)
-        let response = try JSONDecoder().decode(Resource.self, from: data)
-        try fileHandler.saveFile(directoryName: charaDomain, fileName: "resource.json", data: data)
-        return response
-    }
-    
+//    static func downloadResourceJson(charaDomain: String) async throws -> Resource {
+//        let path = "/\(charaDomain)/resource.json"
+//        let url = URL(string: environmentVariable.apiEndpoint + path)!
+//        let data = try await APIClient<String>().fileDownload(url: url)
+//        let response = try JSONDecoder().decode(Resource.self, from: data)
+//        try fileHandler.saveFile(directoryName: charaDomain, fileName: "resource.json", data: data)
+//        return response
+//    }
+//    
     
     static func downloadResource(charaDomain: String, directory: String, fileName: String) async throws {
         let path = "\(environmentVariable.resourceEndpoint)/\(charaDomain)/\(directory)/\(fileName)"
