+++ CharacterAlarm/UseCase/CharaResourceUseCase.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+struct CharaResourceUseCase {
+    private let fileRepository = FileRepository()
+    
+    
+}
